review: regex tests moved to their own file
